Final Patch

+++ controllers/blogs.js
@@ -6,6 +6,7 @@ const jwt = require("jsonwebtoken");
 // ---------------------------------------Create One Blog---------------------------------- //
 
 exports.createBlog = (req, res, next) => {
+  //Vérification s'il y a un fichier ou non, puis création d'un objet en récupérant les éléments de la requête pour les envoyer par la suite à la BDD.
   if (req.file) {
     const blog = new Blog({
       title: req.body.title,
@@ -46,20 +47,43 @@ exports.createBlog = (req, res, next) => {
 // ---------------------------------------Modify One Blog---------------------------------- //
 
 exports.modifyOneBlog = (req, res, next) => {
+  //Récupération des éléments de la requête (role et userID)
   const reqID = req.body.id;
   const userRole = req.body.userRole;
   console.log(userRole);
 
+  //Récupération du Blog qu'on veut modifier dans la BDD
   Blog.findOne({ _id: reqID }).then((blog) => {
     blogUserId = blog.userId;
 
+    //En fonction du Role (1945 =admin) et si le UserId de la req matche avec le UserId de celui qui a créé le blog, on permet la modification du Blog
     if (userRole == 1945 || blogUserId === req.body.userId) {
+      //s'il y a un fichier, on va supprimer l'ancien fichier, et le remplacer par le nouveau, puis sauvegarder l'object blog dans la BDD
       if (req.file) {
         Blog.findOne({ _id: reqID }).then((blog) => {
-          const oldfilename = blog.imageUrl.split("/images/")[1];
-          console.log(blog.imageUrl);
-          console.log(oldfilename);
-          fs.unlink(`images/${oldfilename}`, () => {
+          //vérification si le blog existant avait déjà une image ou non
+          if (blog.imageUrl) {
+            const oldfilename = blog.imageUrl.split("/images/")[1];
+            console.log(blog.imageUrl);
+            console.log(oldfilename);
+            fs.unlink(`images/${oldfilename}`, () => {
+              Blog.updateOne(
+                { _id: reqID },
+                {
+                  body: req.body.body,
+                  title: req.body.title,
+                  author: req.body.author,
+                  imageUrl: `${req.protocol}://${req.get("host")}/images/${
+                    req.file.filename
+                  }`,
+                }
+              )
+                .then(() => {
+                  res.status(200).json({ message: "Objet modifié !" });
+                })
+                .catch((error) => res.status(400).json({ error }));
+            });
+          } else {
             Blog.updateOne(
               { _id: reqID },
               {
@@ -75,10 +99,10 @@ exports.modifyOneBlog = (req, res, next) => {
                 res.status(200).json({ message: "Objet modifié !" });
               })
               .catch((error) => res.status(400).json({ error }));
-          });
+          }
         });
       } else {
-        console.log("no file here, updating body");
+        console.log("Pas de fichier, updating body");
         Blog.updateOne(
           { _id: reqID },
           {
@@ -118,6 +142,7 @@ exports.getOneBlog = (req, res, next) => {
 // ---------------------------------------Delete Blog---------------------------------- //
 
 exports.deleteOneBlog = (req, res, next) => {
+  //Récupération et stockage des données de l'utilisateur contenues dans les headers de la req
   const token = req.headers.authorization.split(" ")[1];
   const decodedToken = jwt.verify(token, process.env.ACCESS_TOKEN_SECRET);
 
@@ -129,7 +154,9 @@ exports.deleteOneBlog = (req, res, next) => {
   Blog.findOne({ _id: req.params.id }).then((blog) => {
     blogUserId = blog.userId;
 
+    //Protection pour s'assurer que le role ou le ID matche et permet de delete le blog de façon sécuritaire. Ces éléments sont aussi protégés dans le Frontend
     if (userRole == 1945 || blogUserId === reqID) {
+      //s'il y a un fichier, on va supprimer l'ancien fichier, et le remplacer par le nouveau, puis sauvegarder l'object blog dans la BDD
       if (req.file) {
         Blog.findOne({ _id: req.params.id })
           .then((blog) => {
@@ -159,18 +186,20 @@ exports.deleteOneBlog = (req, res, next) => {
 // ---------------------------------------Like + Dislike---------------------------------- //
 
 exports.likeBlog = (req, res, next) => {
-  //Fonctions et Variables
+  //Stockage des données de la REQ dans des variables
   let likeValue = req.body.likeValue;
   let blogId = req.body.blogId;
   let userIdReq = req.body.userId;
 
   console.log(likeValue, blogId, userIdReq);
 
+  //récupération du blog demandé et stockage des données concernant les likes en local
   Blog.findOne({ _id: blogId })
     .then((blog) => {
       likesTable = blog.userLiked;
       console.log(likesTable);
 
+      //en fonction de LikeValue, on demande de rajouter ou d'enlever un like. On va vérifier à chaque fois quel est le statut, et agir en conséquence en rajoutant ou enlevant un like et la mention de l'utilisateur dans le tableau
       if (likeValue === 1) {
         if (!likesTable.includes(userIdReq)) {
           console.log("User did not vote..");
